// apps/admin/src/api/generated/schemas/invoiceRead.ts
export interface InvoiceRead {
  /** Creation timestamp */
  createdAt?: InvoiceReadCreatedAt;
  /** Last update timestamp */
  updatedAt?: InvoiceReadUpdatedAt;
  id?: InvoiceReadId;
  invoiceNumber?: InvoiceReadInvoiceNumber;
  saleId?: InvoiceReadSaleId;
  branchId?: InvoiceReadBranchId;
  partyId?: InvoiceReadPartyId;
  deviceId?: InvoiceReadDeviceId;
  tenantId?: InvoiceReadTenantId;
  deviceName?: InvoiceReadDeviceName;
  deviceSerial?: InvoiceReadDeviceSerial;
  devicePrice?: InvoiceReadDevicePrice;
  partyName?: InvoiceReadPartyName;
  partyTC?: InvoiceReadPartyTC;
  status?: InvoiceReadStatus;
  sentToGib?: InvoiceReadSentToGib;
  sentToGibAt?: InvoiceReadSentToGibAt;
  notes?: InvoiceReadNotes;
  createdBy?: InvoiceReadCreatedBy;
  edocumentStatus?: InvoiceReadEdocumentStatus;
  edocumentType?: InvoiceReadEdocumentType;
  ettn?: InvoiceReadEttn;
  profileId?: InvoiceReadProfileId;
  invoiceTypeCode?: InvoiceReadInvoiceTypeCode;
  hasGibPdf?: InvoiceReadHasGibPdf;
  hasGibXml?: InvoiceReadHasGibXml;
  gibPdfLink?: InvoiceReadGibPdfLink;
  birfaturaSentAt?: InvoiceReadBirfaturaSentAt;
  birfaturaApprovedAt?: InvoiceReadBirfaturaApprovedAt;
  taxOffice?: InvoiceReadTaxOffice;
  returnReferenceNumber?: InvoiceReadReturnReferenceNumber;
  returnReferenceDate?: InvoiceReadReturnReferenceDate;
  remoteMessage?: InvoiceReadRemoteMessage;
  metadata?: InvoiceReadMetadata;
}